+++ components/zcash_protocol/CHANGELOG.md
@@ -13,6 +13,7 @@ workspace.
 ### Added
 - `zcash_protocol::consensus::TxIndex`
 - `zcash_protocol::consensus::COINBASE_MATURITY_BLOCKS`
+- `zcash_protocol::consensus::BranchId::{has_sprout, has_sapling, has_orchard}`
 
 ## [0.7.2] - 2025-12-10